+++ cocos/physics/ammo/ammo-rigid-body.ts
@@ -39,7 +39,7 @@ import { ERigidBodyType } from '../framework/physics-enum';
 import { AmmoSharedBody } from './ammo-shared-body';
 import { IVec3Like } from '../../core/math/type-define';
 import { AmmoConstant, CC_V3_0, CC_V3_1 } from './ammo-const';
-import { bt } from './export-bullet';
+import { bt } from './bullet.asmjs';
 
 const v3_0 = CC_V3_0;
 const v3_1 = CC_V3_1;
