+++ cocos/physics/ammo/ammo-util.ts
@@ -33,7 +33,7 @@
 import { IVec3Like, IQuatLike } from '../../core/math/type-define';
 import { Mesh } from '../../3d';
 import { PrimitiveMode } from '../../core/gfx';
-import { bt } from './export-bullet';
+import { bt } from './bullet.asmjs';
 
 export function cocos2BulletVec3 (out: Bullet.ptr, v: IVec3Like): Bullet.ptr {
     bt.Vec3_set(out, v.x, v.y, v.z);
